+++ jkit-sql/src/main/java/com/alianga/jkit/sql/ast/SqlSubqueryTable.java
@@ -48,5 +48,6 @@ public final class SqlSubqueryTable extends SqlTableSource {
     @Override
     protected void acceptChildren(SqlVisitor visitor) {
         child(visitor, query);
+        children(visitor, columnAliases());
     }
 }